Refactored StarSelector to use InvisibleStar child component
