+++ client/src/components/reviews/ReviewList/ReviewForm/StarSelector/InvisibleStar.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const InvisibleStar = props => {
+  return (
+    <div
+      data-star={props.star}
+      className="emptyStar"
+      onMouseOver={(e) => props.setRating(e.target.getAttribute('data-star'))}
+      onMouseLeave={() => props.setRating(props.clickedRating)}
+      onClick={props.handleClick}
+    />
+  );
+};
+
+export default InvisibleStar;
